+++ changelog.md
@@ -2,6 +2,8 @@
 
 ### [Latest]
 
+- Adding `subtract` Method in `VarVsVar` class [#336](https://github.com/umami-hep/puma/pull/336)
+
 ### [v0.4.10](https://github.com/umami-hep/puma/releases/tag/v0.4.10) (2025/08/07)
 
 - Update atlas-ftag-tools Version to v0.2.15 [#335](https://github.com/umami-hep/puma/pull/335)